README.md
This project shows the basic of user login register module in without worrying about json object. The data is sent to the server and retrieve from the server without json format. It's simple and easy to use this module for your application. You have to just pass you webiste URLs and you have to add the php script(included here) on you're server & that's it try using it.